+++ src/client/src/api/apiInterfaces.ts
@@ -1,3 +1,5 @@
+import { File } from "./file/fileInterfaces";
+
 export class ApiError extends Error {
   status?: number;
 
@@ -138,3 +140,20 @@ export interface Section {
 export interface GeometryFormat {
   csvHeader: string;
 }
+
+export interface Photo {
+  id: number;
+  boreholeId: number;
+  name: string;
+  nameUuid: string;
+  fileType: string;
+  fromDepth: number;
+  toDepth: number;
+  public: boolean;
+  createdBy?: User;
+  created?: Date | string;
+  updatedBy?: User;
+  updated?: Date | string;
+}
+
+export type BoreholeAttachment = File | Photo;