+++ tests/api/Controllers/ExportControllerTest.cs
@@ -61,32 +61,28 @@ public class ExportControllerTest
         boreholeFileCloudServiceLoggerMock.Setup(l => l.Log(It.IsAny<LogLevel>(), It.IsAny<EventId>(), It.IsAny<It.IsAnyType>(), It.IsAny<Exception>(), (Func<It.IsAnyType, Exception, string>)It.IsAny<object>()));
         boreholeFileCloudService = new BoreholeFileCloudService(context, configuration, boreholeFileCloudServiceLoggerMock.Object, contextAccessorMock.Object, s3ClientMock);
 
-        var boreholeLockServiceMock = new Mock<IBoreholeLockService>(MockBehavior.Strict);
-        boreholeLockServiceMock
-            .Setup(x => x.IsBoreholeLockedAsync(It.IsAny<int?>(), It.IsAny<string?>()))
-            .ReturnsAsync(false);
-
-        boreholeLockServiceMock
-            .Setup(x => x.IsUserLackingPermissionsAsync(It.IsAny<int?>(), "sub_viewer"))
+        var boreholePermissionServiceMock = new Mock<IBoreholePermissionService>(MockBehavior.Strict);
+        boreholePermissionServiceMock
+            .Setup(x => x.CanEditBoreholeAsync(It.IsAny<string?>(), It.IsAny<int?>()))
             .ReturnsAsync(true);
 
-        boreholeLockServiceMock
-            .Setup(x => x.IsUserLackingPermissionsAsync(It.IsAny<int?>(), "sub_admin"))
+        boreholePermissionServiceMock
+            .Setup(x => x.CanEditBoreholeAsync("sub_viewer", It.IsAny<int?>()))
             .ReturnsAsync(false);
 
-        boreholeLockServiceMock
-            .Setup(x => x.HasUserWorkgroupPermissions(It.IsAny<Borehole>(), It.IsAny<User>()))
-            .Returns(true);
+        boreholePermissionServiceMock
+            .Setup(x => x.CanEditBoreholeAsync("sub_admin", It.IsAny<int?>()))
+            .ReturnsAsync(true);
 
-        boreholeLockServiceMock
-            .Setup(x => x.IsUserLackingPermissions(It.IsAny<Borehole>(), It.IsAny<User>()))
-            .Returns(false);
+        boreholePermissionServiceMock
+            .Setup(x => x.CanViewBoreholeAsync(It.IsAny<string>(), It.IsAny<int>()))
+            .ReturnsAsync(true);
 
         var boreholeFileControllerLoggerMock = new Mock<ILogger<BoreholeFileController>>(MockBehavior.Strict);
         boreholeFileControllerLoggerMock.Setup(l => l.Log(It.IsAny<LogLevel>(), It.IsAny<EventId>(), It.IsAny<It.IsAnyType>(), It.IsAny<Exception>(), (Func<It.IsAnyType, Exception, string>)It.IsAny<object>()));
 
         var loggerMock = new Mock<ILogger<ExportController>>();
-        controller = new ExportController(context, boreholeFileCloudService, loggerMock.Object, boreholeLockServiceMock.Object) { ControllerContext = GetControllerContextAdmin() };
+        controller = new ExportController(context, boreholeFileCloudService, loggerMock.Object, boreholePermissionServiceMock.Object) { ControllerContext = GetControllerContextAdmin() };
     }
 
     [TestMethod]
@@ -539,19 +535,19 @@ public class ExportControllerTest
     }
 
     [TestMethod]
-    public async Task ExportControllerMethodsShouldValidateUserLacksPermissions()
+    public async Task ExportControllerMethodsShouldValidateUserHasPermissions()
     {
         // Override return value of HasUserWorkgroupPermissions in this specific test
-        var boreholeLockServiceMock = new Mock<IBoreholeLockService>(MockBehavior.Loose);
-        boreholeLockServiceMock
-            .Setup(x => x.IsUserLackingPermissions(It.IsAny<Borehole>(), It.IsAny<User>()))
-            .Returns(true);
+        var boreholePermissionServiceMock = new Mock<IBoreholePermissionService>(MockBehavior.Loose);
+        boreholePermissionServiceMock
+            .Setup(x => x.CanViewBoreholeAsync(It.IsAny<string>(), It.IsAny<int>()))
+            .ReturnsAsync(false);
 
         var boreholeFileControllerLoggerMock = new Mock<ILogger<BoreholeFileController>>(MockBehavior.Strict);
         boreholeFileControllerLoggerMock.Setup(l => l.Log(It.IsAny<LogLevel>(), It.IsAny<EventId>(), It.IsAny<It.IsAnyType>(), It.IsAny<Exception>(), (Func<It.IsAnyType, Exception, string>)It.IsAny<object>()));
 
         var loggerMock = new Mock<ILogger<ExportController>>();
-        controller = new ExportController(context, boreholeFileCloudService, loggerMock.Object, boreholeLockServiceMock.Object) { ControllerContext = GetControllerContextAdmin() };
+        controller = new ExportController(context, boreholeFileCloudService, loggerMock.Object, boreholePermissionServiceMock.Object) { ControllerContext = GetControllerContextAdmin() };
 
         // Get all export methods from ExportController
         var exportMethods = typeof(ExportController)