+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+### Added
+- Added the photos tab to the borehole attachments.
+- Added photo viewer to the borehole sidepanel.
+
+### Fixed
+- Boreholes were always copied to the default workgroup instead of the selected workgroup.
+
+## v2.1.1162 - 2025-04-10
+
 ### Fixed
 - Calculations of TVD resulted in an error when the distance between the two coordinates to interpolate was larger than their difference of MD.